+++ app/Models/ValueObjects/MenuItem.php
@@ -11,6 +11,7 @@ use Kami\Cocktail\Models\Enums\MenuItemTypeEnum;
 readonly class MenuItem
 {
     private function __construct(
+        public int $id,
         public MenuItemTypeEnum $type,
         public int $sort,
         public Price $price,
@@ -25,6 +26,7 @@ readonly class MenuItem
     public static function fromMenuCocktail(MenuCocktail $menuCocktail): self
     {
         return new self(
+            id: $menuCocktail->cocktail_id,
             type: MenuItemTypeEnum::Cocktail,
             sort: $menuCocktail->sort,
             price: new Price($menuCocktail->getMoney()),
@@ -39,12 +41,13 @@ readonly class MenuItem
     public static function fromMenuIngredient(MenuIngredient $menuIngredient): self
     {
         return new self(
+            id: $menuIngredient->ingredient_id,
             type: MenuItemTypeEnum::Ingredient,
             sort: $menuIngredient->sort,
             price: new Price($menuIngredient->getMoney()),
             name: $menuIngredient->ingredient->name,
             categoryName: $menuIngredient->category_name,
-            description: $menuIngredient->ingredient->description,
+            description: $menuIngredient->ingredient->getMaterializedPathAsString(),
             publicId: null,
             image: config('app.url') . $menuIngredient->ingredient->getMainImageThumbUrl(false),
         );